src/python/functions/func_dataproc/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-functions-durable
azure-storage-blob
azure-core
opencensus-extension-azure-functions
